+++ server/app.js
@@ -15,8 +15,7 @@ const app = express({
 });
 
 app.get('/__gtg', controllers.gtg);
-app.get('/videos', controllers.video);
-app.get('/videos/:id', controllers.video);
+app.get('/videos', controllers.hub);
 
 const listen = app.listen(process.env.PORT || 3001);
